corrects

+++ lesson_01/squares.rb
@@ -1,21 +1,21 @@
 puts "a?"
-a = gets.chomp.to_i
+a = gets.chomp.to_f
 
 puts "b?"
-b  = gets.chomp.to_i
+b  = gets.chomp.to_f
 
 puts "c?"
-c = gets.chomp.to_i
+c = gets.chomp.to_f
 
 d = b**2 - 4*a*c
 
 if d<0
   puts "корней нет"
 elsif d==0
-  x = -b/(2.0*a)
+  x = -b/(2*a)
   puts "d = #{d}, один корень x = #{x}"
 else
-  x1 = (-b + Math.sqrt(d))/(2.0*a)
-  x2 = (-b - Math.sqrt(d))/(2.0*a)
+  x1 = (-b + Math.sqrt(d))/(2*a)
+  x2 = (-b - Math.sqrt(d))/(2*a)
   puts "d = #{d}, x1 = #{x1}, x2 = #{x2}"
 end
